ui: optimise the code and UI for value view
